fix test naming + add blocking specific test
